+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   * Improved enrollment for HOTP and TOTP tokens
   * Removed non-public dependencies. LinOTP Selfservice can now be build by everyone.
   * Dynamically change the theme based on the device's color scheme, without requiring a browser refresh.
+  * Display warnings about not verified tokens
 
   ### Dependencies
   * Update to Angular 16